Delete DirectedEdgeNLevelSparseVisibilityGraphAlgorithm. Will be replaced with another algorithm of the same name later.
I had originally thought that making the edges directed would reduce the amount of nodes searched. My intuition was correct, but I had formalised it the wrong way.
This isn't the way to do the directed edge version.
Thus the algorithm has been deleted, and I will soon replace it with a more proper formalism of the Directed ENLSVG algorithm.